+++ src/ladb_opencutlist/ruby/utils/price_utils.rb
@@ -18,6 +18,7 @@ module Ladb::OpenCutList
     def fetch_currency_options
       settings_model = Plugin.instance.get_model_preset('settings_model')
       @currency_symbol = settings_model['currency_symbol']
+      @currency_precision = settings_model['currency_precision'].to_i
     end
 
     def get_symbol
@@ -29,8 +30,8 @@ module Ladb::OpenCutList
     # Take a float containing a price
     # and convert it to a string representation according to the
     # local unit settings.
-    def format_to_readable_price(f, precision = 0)
-      UnitUtils.format_readable(f, UNIT_STRIPPEDNAME, [ 2, precision ].min, 2)
+    def format_to_readable_price(f)
+      UnitUtils.format_readable(f, UNIT_STRIPPEDNAME, @currency_precision, [ 2, @currency_precision ].max)
     end
 
   end