+++ CHANGELOG.md
@@ -16,6 +16,7 @@ in 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
   * Added export remove all columns button
   * Added table row size options
   * Added Toggle button in cutlist tags filters to switch from "present" to "not present"
+  * Added Mass and Currency custom precision in settings tab
   * Changed material color bullet from drop to circle
 
 * 5.0.4 (2023-09-14)